pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gra-forum"
version = "0.1.0"
description = "AI Agent Debate Orchestrator based on GRA principles"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "GRA Forum Team", email = "contact@gra-forum.org"}
]
keywords = ["ai", "agents", "debate", "orchestration", "gra", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.8.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-asyncio>=0.21.0", "black>=23.0.0", "mypy>=1.5.0"]
embeddings = ["sentence-transformers>=2.2.0"]

[project.urls]
Homepage = "https://github.com/yourusername/gra-forum"
Documentation = "https://gra-forum.readthedocs.io"
Repository = "https://github.com/yourusername/gra-forum"
Issues = "https://github.com/yourusername/gra-forum/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["gra_forum*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
